i'm apparently confused about rules for math and unsigned ints
git-svn-id: https://svn.salilab.org/imp/trunk@7878 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/em2d/src/image_processing.cpp
@@ -523,7 +523,8 @@ void do_dilate_and_shrink_warp(algebra::Matrix2D_d &m,
 //    xxx.set_data(temp);
 //    xxx.write("temp.spi",srw);
     /*******/
-  } while( std::abs(static_cast<int>(new_size_in_pixels)-size_in_pixels)>1);
+  } while( std::abs(static_cast<int>(new_size_in_pixels)
+                    -static_cast<int>(size_in_pixels))>1);
 //  } while( new_size_in_pixels != size_in_pixels);
   m.copy(temp);
 }
